Move files from data/model to appropriate folders
