+++ src/common/src/main/kotlin/org/calypsonet/keyple/demo/common/model/ReaderType.kt
@@ -1,5 +1,5 @@
 /* ******************************************************************************
- * Copyright (c) 2021 Calypso Networks Association https://calypsonet.org/
+ * Copyright (c) 2022 Calypso Networks Association https://calypsonet.org/
  *
  * See the NOTICE file(s) distributed with this work for additional information
  * regarding copyright ownership.
@@ -10,8 +10,11 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  ****************************************************************************** */
-package org.calypsonet.keyple.demo.reload.remote.data.model
+package org.calypsonet.keyple.demo.control.data.model
 
-object AppSettings {
-  var aidEnums = arrayListOf<ByteArray>()
+enum class ReaderType {
+  BLUEBIRD,
+  COPPERNIC,
+  FAMOCO,
+  NFC_TERMINAL
 }
