+++ src/reloading-remote/client/keyple-mobile-android/app/src/main/kotlin/org/calypsonet/keyple/demo/reload/remote/domain/model/AppSettings.kt
@@ -1,0 +1,5 @@
+package org.calypsonet.keyple.demo.reload.remote.domain.model
+
+object AppSettings {
+  var aidEnums = arrayListOf<ByteArray>()
+}
