+++ src/reloading-remote/client/keyple-mobile-android/app/src/main/kotlin/org/calypsonet/keyple/demo/reload/remote/domain/model/ServerStatusEvent.kt
@@ -1,0 +1,3 @@
+package org.calypsonet.keyple.demo.reload.remote.domain.model
+
+class ServerStatusEvent(val isUp: Boolean)
